correct the iteration to jump into and do the bitmasking by casting to uint16_t

+++ src/core/Math.hpp
@@ -8,6 +8,6 @@ struct Vec3fx { fx x, y, z; };
 struct Vec2i  { int16_t x, y; };
 
 // Integer square root of a 32-bit unsigned value (floor).
-uint32_t isqrt32(uint32_t x);
+uint16_t isqrt32(uint32_t x);
 
 } // namespace gv
